refactor: rename routes folder to pages
