+++ apps/web/src/router.tsx
@@ -1,10 +1,10 @@
 import { BrowserRouter, Route, Routes } from "react-router";
 import Layout from "./layouts/Layout";
-import Home from "./routes/Home";
-import Login from "./routes/Login";
-import Register from "./routes/Register";
+import Home from "./pages/Home";
+import Login from "./pages/Login";
+import Register from "./pages/Register";
 import { AuthLayout } from "./layouts/AuthLayout";
-import Todos from "./routes/Todos";
+import Todos from "./pages/Todos";
 
 export default function Router() {
   return (
